+++ Cargo.toml
@@ -4,6 +4,12 @@ version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path ="src/zos/mod.rs"
+
+[[bin]]
+path = "src/bin/main.rs"
+name = "zos"
 
 [dependencies]
 rbus = {path="../rbus"}
