+++ src/bin/main.rs
@@ -1,9 +1,10 @@
+mod modules;
+
 use clap_v3::App;
 // use rbus;
 // use std::time::Duration;
-mod zos;
-use zos::modules::zui::zui::run;
 use std::error::Error;
+
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
     let matches = App::new("Zero-OS")
@@ -13,15 +14,15 @@ async fn main() -> Result<(), Box<dyn Error>> {
             App::new("zui")
                 .about("Show Zero os UI")
                 .version("1.0")
-                
+
         )
         .get_matches();
 
     match matches.subcommand() {
-        ("zui", Some(_sub_m)) => run().await?,
+        ("zui", Some(_sub_m)) => modules::zui::run().await?,
         _ => {
             println!("Welcome to zos, please supply subcommand or --help or more info")
-        } 
+        }
     }
     Ok(())
 }
