+++ src/zos/mod.rs
@@ -1,4 +1,4 @@
 pub mod app;
 pub mod bus;
-pub mod modules;
-pub mod pkg;
+pub mod environment;
+pub mod kernel;
