--- src/zos/modules/zui/mod.rs
@@ -1,3 +1,0 @@
-mod app;
-mod ui;
-pub mod zui;
