--- src/zos/pkg/environment/mod.rs
@@ -1,1 +1,0 @@
-pub mod environment;
