--- src/zos/pkg/kernel/mod.rs
@@ -1,1 +1,0 @@
-pub mod kernel;
