--- src/zos/pkg/mod.rs
@@ -1,2 +1,0 @@
-pub mod environment;
-pub mod kernel;
